commit spacemacs.el file; shouldn't be edited but is informative
THe spacemacs.el file should not be edited itself, but rather the org file
should be edited and then spacemacs.el should be regenerated. I removed it from
the repo earlier for that reason, but now I'm adding it back. The reason is that
it might help others to see the result of the tangling, or if someone who
doesn't use a literate config and wants to grab from my dotfile this might be
easier.